core: move test listener to root package
