+++ webtau-junit5/src/main/java/org/testingisdocumenting/webtau/junit5/WebTauJunitExtension.java
@@ -16,6 +16,7 @@
 
 package org.testingisdocumenting.webtau.junit5;
 
+import org.testingisdocumenting.webtau.TestListeners;
 import org.testingisdocumenting.webtau.javarunner.report.JavaBasedTest;
 import org.testingisdocumenting.webtau.javarunner.report.JavaReport;
 import org.testingisdocumenting.webtau.javarunner.report.JavaShutdownHook;
