+++ webtau-standalone-runner-groovy/src/main/groovy/org/testingisdocumenting/webtau/runner/standalone/StandaloneTest.groovy
@@ -17,7 +17,7 @@
 
 package org.testingisdocumenting.webtau.runner.standalone
 
-import org.testingisdocumenting.webtau.reporter.TestListeners
+import org.testingisdocumenting.webtau.TestListeners
 import org.testingisdocumenting.webtau.reporter.WebTauTest
 import org.testingisdocumenting.webtau.reporter.StepReporter
 import org.testingisdocumenting.webtau.reporter.StepReporters
